Remove component_interface_value_fixtures.py
